organized mock implementations into a dedicated mocks directory
